Move Profiler.updated to compile-common
