+++ packages/compile-common/src/profiler/index.ts
@@ -6,3 +6,4 @@ export { isExplicitlyRelative } from "./isExplicitlyRelative";
 export { minimumUpdatedTimePerSource } from "./minimumUpdatedTimePerSource";
 export { resolveAllSources } from "./resolveAllSources";
 export { requiredSources } from "./requiredSources";
+export { updated } from "./updated";
